Move CLI source code to its own directory
Let's separate the CLI part, which provisions the
testing runtime, from the real POC for metricbeat. This eventually will
help when adding other POCS, i.e. the Kibana one
