+++ cli/go.mod
@@ -1,0 +1,36 @@
+module github.com/elastic/metricbeat-tests-poc
+
+go 1.12
+
+require (
+	cloud.google.com/go v0.43.0 // indirect
+	github.com/DATA-DOG/godog v0.7.13
+	github.com/Microsoft/go-winio v0.4.12 // indirect
+	github.com/docker/distribution v2.7.1+incompatible // indirect
+	github.com/docker/docker v0.7.3-0.20190506211059-b20a14b54661
+	github.com/docker/go-connections v0.4.0
+	github.com/docker/go-units v0.4.0 // indirect
+	github.com/gorilla/mux v1.7.3 // indirect
+	github.com/kisielk/errcheck v1.2.0 // indirect
+	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
+	github.com/kr/pty v1.1.8 // indirect
+	github.com/logrusorgru/aurora v0.0.0-20190428105938-cea283e61946
+	github.com/mitchellh/mapstructure v1.1.2
+	github.com/pkg/errors v0.8.1 // indirect
+	github.com/sirupsen/logrus v1.4.2 // indirect
+	github.com/spf13/cobra v0.0.5
+	github.com/spf13/viper v1.3.2
+	github.com/stretchr/objx v0.2.0 // indirect
+	github.com/stretchr/testify v1.3.0
+	github.com/testcontainers/testcontainers-go v0.0.5-0.20190614112146-b130c69aad64
+	golang.org/x/crypto v0.0.0-20190701094942-4def268fd1a4 // indirect
+	golang.org/x/exp v0.0.0-20190718202018-cfdd5522f6f6 // indirect
+	golang.org/x/image v0.0.0-20190703141733-d6a02ce849c9 // indirect
+	golang.org/x/mobile v0.0.0-20190719004257-d2bd2a29d028 // indirect
+	golang.org/x/net v0.0.0-20190628185345-da137c7871d7 // indirect
+	golang.org/x/sys v0.0.0-20190712062909-fae7ac547cb7 // indirect
+	golang.org/x/tools v0.0.0-20190719005602-e377ae9d6386 // indirect
+	google.golang.org/grpc v1.22.0 // indirect
+)
+
+replace github.com/spf13/viper v1.3.2 => github.com/mdelapenya/viper v1.4.1
